move preview popover to space plugin
